# apps/Gec_Server_C/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GECChain MCP API (Aptos Move)",
    "version": "1.0.0",
    "description": "MVP OpenAPI spec aligned with the current Move modules: gec_certificate, gec_marketplace, gec_audit. Intended for MCP tool integration (agent -> API -> Aptos tx)."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local backend API"
    },
    {
      "url": "https://YOUR_DOMAIN",
      "description": "Deployed backend API"
    }
  ],
...
  "paths": {
...
    "/marketplace/list": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_list",
        "summary": "List a certificate for sale",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/cancel": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_cancel",
        "summary": "Cancel a listing (seller-only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelListingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/request-buy": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_request_buy",
        "summary": "Request to buy a listed certificate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBuyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/accept-buy": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_accept_buy",
        "summary": "Seller accepts a pending buy request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptBuyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/{market_addr}/stats": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_stats",
        "summary": "Marketplace stats via view functions",
        "parameters": [
          {
            "name": "market_addr",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceStats"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TxRequest": {
        "type": "object",
        "properties": {
          "sender_private_key_hex": {
            "type": "string",
            "description": "Ed25519 private key hex for signing (DO NOT expose in production; use wallet or server key vault)."
          },
          "sender_address": {
            "type": "string",
            "description": "0x... Aptos account address"
          },
          "gas_unit_price": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          },
          "max_gas_amount": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          }
        },
        "required": [
          "sender_private_key_hex",
          "sender_address"
        ]
      },
      "TxResponse": {
        "type": "object",
        "properties": {
          "tx_hash": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "vm_status": {
            "type": "string",
            "nullable": true
          },
          "explorer_url": {
            "type": "string",
            "nullable": true
          },
          "sender_address": {
            "type": "string",
            "nullable": true
          },
          "account_id": {
            "type": "string",
            "nullable": true
          },
          "owner_account_id": {
            "type": "string",
            "nullable": true
          },
          "source_account_id": {
            "type": "string",
            "nullable": true
          },
          "target_account_id": {
            "type": "string",
            "nullable": true
          },
          "issued_quantity": {
            "type": "integer",
            "nullable": true
          },
          "transferred_quantity": {
            "type": "integer",
            "nullable": true
          },
          "retired_quantity": {
            "type": "integer",
            "nullable": true
          },
          "energy_source": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "cert_id": {
            "type": "integer",
            "nullable": true
          },
          "listing_id": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "nullable": true
          },
          "module_address": {
            "type": "string",
            "nullable": true
          },
          "display_id": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "tx_hash",
          "success"
        ]
      },
...
      "ListCertificateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "market_addr": {
            "type": "string",
            "description": "Address where MarketplaceStore resource lives"
          },
          "cert_id": {
            "type": "integer",
            "minimum": 1
          },
          "price": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "market_addr",
          "cert_id",
          "price"
        ]
      },
      "CancelListingRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "market_addr": {
            "type": "string"
          },
          "listing_id": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "market_addr",
          "listing_id"
        ]
      },
      "RequestBuyRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "market_addr": {
            "type": "string"
          },
          "listing_id": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "market_addr",
          "listing_id"
        ]
      },
      "AcceptBuyRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "market_addr": {
            "type": "string"
          },
          "listing_id": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "market_addr",
          "listing_id"
        ]
      },
      "MarketplaceStats": {
        "type": "object",
        "properties": {
          "listing_count": {
            "type": "integer"
          },
          "total_trades": {
            "type": "integer"
          },
          "total_volume": {
            "type": "integer"
          }
        },
        "required": [
          "listing_count",
          "total_trades",
          "total_volume"
        ]
      },
...
    }
  }
}